+++ scripts/prepdocslib/files/embeddings.py
