+++ scripts/prepdocslib/strategy.py
